Prevent possible code injection
Interpolating GitHub templating into shell scripts poses the risk of
inadvertent injection of user-provided content.

Although the content is not really untrusted in this example (because
users must have write access already before they trigger this workflow),
it’s still good practice to make this safe in case the security boundary
ever changes, or if someone copies and pastes this.